requirements.txt
beautifulsoup4
jsonlines
numpy
openai==0.27.2
PyMuPDF
python_docx
requests
tqdm
validators
tiktoken
tenacity